API Reference

Some functions in the package serve a general purpose, while others are keyed to data types.

General-purpose functions

There are many such functions. For example, get_file is a cached download function that will download a remote file and install it locally, skipping the download if there is already a local version that was downloaded sufficiently recently. Within the package, this is used for Argo files, but users are free to use it for any other file type. Similarly, pretty provides a way to get simple increments to be used in contouring, but users may find it helpful for a variety of other purposes. In addition to these, there are many functions that deal with water properties, distances between points on the earth, etc. The names are designed to be reasonably straightforward, and so scanning the full API list will be a good guide for users.

Special-purpose functions

The package offers several special-purpose functions that are tied to data types. The following listing provides a rough guideline, keyed to the data type. The examples should also be helpful in getting acquainted with the functions, and the data types.

Type inheritance

OA is the base type in the package. All of the following types inherit from OA.

Ctd holds hydrographic data, from plain CTD instruments or Argo files. See as_ctd, read_ctd_cnv, and read_argo for constructing data objects of type `Ctd. See plot_profile and plot_TS for plotting them. .

Amsr holds data from AMSR satellites. See read_amsr for reading such files and plotting their contents.

Coastline holds coastline shapes. See coastline for reading and plotting such data.

Links to detailed type documentation

OceanAnalysis.OA — Type

Base type in the OceanAnalysis package.

This is an abstract type. The other types in the package will derive from this. At the moment, the only such case is Ctd.

OceanAnalysis.Amsr — Type

A type to hold AMSR data (SUBJECT TO CHANGE)

This holds AMSR satellite data as read by read_amsr.

The metadata element is a Dict that holds the source filename, the longitude and latitude, and the name of the stored variable.

The data element stores the data.

OceanAnalysis.Ctd — Type

A type to hold CTD data

Ctd is a type used to store data from CTD instruments and Argo floats. It takes the form of a Struct that derives from the base OA type, and holds two items: (1) a DataFrame named data that holds the actual data, including pressure, salinity, and temperature, perhaps along with other data and (2) a a Dict named metadata that stores information about the data, such as the time of observation and the latitude and longitude at which the observation was made.

Objects of type Ctd are returned by as_ctd, read_ctd_cnv and read_argo and can be passed to plotting functions plot_profile and plot_TS, and by several functions relating to seawater properties, such as SA and other TEOS-10 related functions, as well as functions relating to the distributions of such properties, such as N2.

OceanAnalysis.Coastline — Type

A type to hold coastline data

Coastline is a type to hold coastline data. Its metadata element is a Dict that may hold the source filename or other information (or may be empty). Its data element is a DataFrame holding columns named longitude and latitude, with NaN values separating continents and/or elements within them such as countries or subregions of countries.

Links to detailed function documentation

OceanAnalysis.argo_id_cycle — Function

Split Argo "id_cycle" into components id and cycle

Examples

julia> using OceanAnalysis

julia> argo_id_cycle("4902911_095")
2-element Vector{SubString{String}}:
 "4902911"
 "095"

OceanAnalysis.as_ctd — Function

as_ctd(salinity, temperature, pressure, longitude=NaN, latitude=NaN; time, debug=0)

Construct a Ctd object, given S, T, p, and a location.

Returns a Ctd object with a data element that is a data frame holding the provided water properties, along with computed Absolute Salinity (SA) Conservative Temperature (CT), potential density anomaly relative to the surface pressure (sigma0) and spiciness with respect to surface pressure (spiciness0). The object also holds a metadata element that holds longitude, latitude and time. If either longitude or latitude is NaN, thenSA, etc. are computed assuming a mid-Atlantic location (-30E and 30N).

Arguments

salinity::Vector{Float64} measured salinity values, in Practical Salinity units.

temperature::Vector{Float64} measured temperature values, in degrees Celsius.

pressure::Vector{Float64} measured sea pressure, in dbar.

longitude::Float64 observation longitude, in degrees East. If not provided, this defaults to -30 (i.e. -30E, or 30W, in the North Atlantic).

latitude::Float64 observation latitude, in degrees North. If not provided, this defaults to 30 (i.e. 30N, in the North Atlantic).

time::Date.DateTime an optional indication of the measurement start time.

add_teos::Bool an optional indication of whether to add SA, CT, sigma0 and spiciness0 to the data component of the return value.

debug::Int64 an optional value that, if it exceeds 0, indicates that debugging output should be printed during processing.

Examples

julia> using OceanAnalysis

julia> as_ctd([32.], [15.], [0.], -63., 40.)
Ctd(Dict{String, Any}("latitude" => 40.0, "time" => nothing, "longitude" => -63.0), 1×7 DataFrame
 Row │ salinity  temperature  pressure  SA       CT       sigma0   spiciness0
     │ Float64   Float64      Float64   Float64  Float64  Float64  Float64
─────┼────────────────────────────────────────────────────────────────────────
   1 │     32.0         15.0       0.0  32.1516  15.0642  23.6653   0.0686905)

julia> as_ctd([32.], [15.], [0.], -63., 40., add_teos=false)
Ctd(Dict{String, Any}("latitude" => 40.0, "time" => nothing, "longitude" => -63.0), 1×3 DataFrame
 Row │ salinity  temperature  pressure
     │ Float64   Float64      Float64
─────┼─────────────────────────────────
   1 │     32.0         15.0       0.0)

OceanAnalysis.coastline — Function

coastline(symbol::Symbol=:world_fine)

Access a built-in coastline dataset. The only valid choices for name are :world_coarse and :world_fine. These are handled by reading the built-in datasets data/coastline_coarse.csv.gz and datasets data/coastline_fine.csv.gz, respectively.

Examples

# Nova Scotia
using OceanAnalysis, Plots
cl = coastline(:global_fine);
plot_coastline(cl, xlims=(-68, -58), ylims=(43, 48))

coastline(filename::String, header::Integer=0)

Return a coastline stored in the named CSV file (in either text or gzipped form).

The work is done by passing filename and header to CSV.read(). The file must have 1 or more header lines, the last of which must contain column names longitude and latitude. NaN values will be taken to indicate breaks between segments that trace continents, nations, etc.

Examples

# World view
using OceanAnalysis, Plots
dir = dirname(dirname(pathof(OceanAnalysis)))
file = joinpath(dir, "data", "coastline_coarse.csv.gz")
cl = coastline(file, 1)
plot_coastline(cl)

coastline(longitude::Vector{Real}, latitude::Vector{Real})

Create a Coastline from longitude and latitude values. Use NaN values for each of these to indicate breaks in the coastline from continent to continent, nation to nation, etc.

Examples

# Nova Scotia
using OceanAnalysis, CSV, Plots, DataFrames
dir = dirname(dirname(pathof(OceanAnalysis)));
file = joinpath(dir, "data", "coastline_fine.csv.gz");
data = CSV.read(file, DataFrame, header=1);
cl = coastline(data.longitude, data.latitude);
plot_coastline(cl, xlims=(-68, -58), ylims=(43, 48))

OceanAnalysis.coordinate_from_string — Function

degree = coordinate_from_string(s::String)

Try to extract a longitude or latitude from a string. If there are two (space-separated) tokens in the string, the first is taken as the decimal degrees, and the second as decimal minutes. The goal is to parse hand-entered strings, which might contain letters like "W" and "S" (or the same in lower case) to indicate the hemisphere. Humans are quite good at writing confusing strings, so this function is not always helpful.

Examples

julia> using OceanAnalysis

julia> coordinate_from_string("1.5")
1.5

julia> coordinate_from_string("1 30")
1.5

julia> coordinate_from_string("1S 30")
-1.5

julia> coordinate_from_string("27* 14.072 N")
27.234533333333335

julia> coordinate_from_string("111* 31.440 W")
-111.524

OceanAnalysis.CT — Function

CT(SA, temperature, pressure)
CT(ctd)

Compute Conservative Temperature (CT), using gsw_ct_from_t() in the GibbsSeaWater package.

The first form takes single values and returns a single value. The second form extracts values from a Ctd object and then calls the first form as CT.() so that it returns a vector of CT values.

Examples

julia> using OceanAnalysis

julia> CT(35.0, 10.0, 100.0)
9.981322531922249

OceanAnalysis.depth_from_pressure — Function

Compute seawater depth (m) from sea pressure (dbar)

See also z_from_pressure.

Examples

julia> using OceanAnalysis

julia> depth_from_pressure(100.0)
99.16434938694897

OceanAnalysis.fix_gsw_bad_code — Function

Change GSW 'missing' values (9.e15) to NaN

A copy is returned, with x unaltered. See fix_gsw_bad_code! for an in-place version.

OceanAnalysis.fix_gsw_bad_code! — Function

In-place change GSW 'missing' values (9.e15) to NaN

This alters x. See fix_gsw_bad_code for a version that does not alter x.

OceanAnalysis.geod_distance — Function

geod_distance(lon1, lat1, lon2, lat2; a=6378137.00, f=1.0 / 298.257223563)

Return the geodetic distance between two points on earth, in kilometres.

This uses the Vincenty [1975] formulation. According to tests described in that paper, the results can be expected to be accurate to 0.01 mm.

Note that geod_distance() mimics the geodDist() function in the oce R package (Ref. 2), and is based on the R and C++ code used in the latter. The code trail is 5 decades long, starting with a CDC-6600 Fortran version written in 1974 by L. Pfeifer, which was modified for IBM 360 by J. G. Gergen. In 2003, D. Gillis wrote an R version that was modified for inclusion in the oce package by Dan Kelley.

Arguments

lon1 and lat1, in degrees, give the location of a given point on the earth

lon2 and lat2, in degrees, give the location of a given point on the earth

a and f are the semi-major axis and flattening parameter of the reference

ellipsoid. (The default values are highly recommended.)

References

Vincenty,T. 1975. Direct and inverse solutions of geodesics on the ellipsoid with application of nested equations. Survey Review 23(176):88-94. https://www.ngs.noaa.gov/PUBS_LIB/inverse.pdf

Kelley, Dan E., Clark Richards, and Chantelle Layton. “Oce: An R Package for Oceanographic Analysis.” Journal of Open Source Software 7, no. 71 (2022). https://doi.org/10.21105/joss.03594

OceanAnalysis.get_argo_index — Function

get_argo_index(destdir::String="."; age::Real=1.0,
    server::String="https://data-argo.ifremer.fr", debug::Int64=0)

Download an Argo index file, unless an existing local copy is newly downloaded.

The file is obtained from server and stored in destdir, but only if an existing version of the file in destdir is under age days old. The default is to cache index files for one day.

The default destdir is the local directory, but it is common to set this value to some central location so the file can be used by Julia sessions running in multiple directories. (The author sets destdir="~/data/argo", for example.)

Return value

read_argo_index returns the full name of the downloaded (or cached) file.

Use read_argo_index to interpret the downloaded file.

OceanAnalysis.get_argo_file — Function

get_argo_file(file::String="", destdir::String="."; age::Real=1.0; server::String="https://data-argo.ifremer.fr", debug=0)

Download an Argo profile file, if an existing copy is less than age days old.

Arguments

file::String path to the file on a server, of the form

agency/#/profiles/-#_##.nc, where # is the ID number of the float, - tells the status of the file (R for realtime files, D for delayed-mode files, etc.) and ## is an identifier for the cast (usually but not always an integer value). This system matches the file information stored on argo servers, as downloaded by get_argo_index and read by read_argo_index.

destdir::String name of the directory into which to save the file.

age::Real file-caching time in days. If the requested file does not exist

locally then age is ignored and the file is downloaded. It will also be downloaded if there is an existing file but it was last downloaded more than age days ago

Returns

::String: Full name of the local file after downloading, or as cached recently.

Example

# Get most last-named file in the Argo index and plot a temperature profile
using OceanAnalysis
index_file = get_argo_index()
index = read_argo_index(index_file)
argo_file = get_argo_file(index.file[end])
argo = read_argo(argo_file)
plot_profile(argo, which="CT")

OceanAnalysis.get_element — Function

get_element(ctd::Ctd, name::String; debug)

Get an element from an object.

OceanAnalysis.get_file — Function

Download a remote file or identify an existing version if it is young

The contents of url are downloaded and stored as file, but only if either file does not exist locally or its age is less than age days.

OceanAnalysis.get_nc_value — Function

Transform an item from a NetCDF file into a more useable object

Any missing values are converted to NaN. This function is designed only for numerical values, not strings or other types.

Arguments

nc a value returned by NCDataset().

name the name of an object contained in nc.

require_value boolean value indicating whether to report an error

if the desired element consists entirely of bad values.

OceanAnalysis.N2 — Function

N2(ctd::Ctd, s::Float64=0.15; debug::Int64=0)

Compute the square of the buoyancy frequency, N², for a Ctd object. The value is inferred from a smoothing cubic spline that models the dependence of sigma0 on pressure.

In the present version, the spline is fitted with the Dierckx::Spline1D() function (Reference 1), which is provided with equal weights, w, for all points, with k=3 to set the polynomial order to cubic, and with bc="nearest" to control what happens near boundaries. The user has no control over these things, although this might change in a future version of N2().

The user's control rests in s, a smoothing parameter that is passed to Dierckx:Spline1D(). If not specified by the user, this defaults to a value that yields N² curves that are similar to those computed with a default call to swN2() in the oce R package. Users may elect to use larger s values for smoother curves, or smaller ones to get more detail. It would be a mistake not to pair experiments with s values with plots. As a start, it might be useful to examine Reference 2, which compares the R and Julia results.

References

https://github.com/JuliaMath/Dierckx.jl

https://github.com/dankelley/OceanAnalysis.jl/issues/13

Examples

# Demonstrate N2()
using OceanAnalysis, Plots
pkgdir = dirname(dirname(pathof(OceanAnalysis)))
filename = joinpath(pkgdir, "data", "ctd.cnv")
ctd = read_ctd_cnv(filename);
# Basic plot (left), profile-plot in oceanographic convention (right)
p1=plot(N2(ctd), z_from_pressure.(ctd.data.pressure))
p2=plot_profile(ctd, which="N2")
plot(p1, p2)

OceanAnalysis.OceanAnalysis — Module

The OceanAnalysis module is intended to help with the analysis of oceanographic data. It is in a preliminary form, providing help with only four file types: (1) Argo NetCDF files, (2) CTD files in the Seabird CNV format, (3) AMSR satellite NetCDF files and (4) coastline files. In each case, the capabilities are quite limited, reflecting the early stage of the package. Users who need more powerful tools for reading and analysing oceanographic data, ought consider using the oce package in the R language, which has more capabilities and has been in wide use for over a decade.

The functions that read data return objects that are structs holding two items. The first is a Dict named metadata, which has contents that vary with data type. The second is named data, the contents of which depend on the data type. The elements of both may be accessed directly using the dot notation, with for example ctd.data.salinity refering to the salinity column of ctd data read by read_ctd_cnv. It is also possible to retrieve (or set) that value using ctd["salinity"]. The advantage of this notation is that it can locate information whether it is in the data or the metadata component of the object. It is also possible to obtain derived information, e.g. ctd["SA"] calculates and then returns the Absolute Salinity, which is not typically stored in CTD files, but which can be computed from the stored information. Other derivable items include "CT" (Conservative Temperature), "sigma0" (potential density anomaly with respect to surface pressure), and spiciness0 (spiciness).

OceanAnalysis.plot_coastline — Function

plot_coastline(coastline::Coastline;
    xlims=(-180., 180.), ylims=(-90., 90.),
    seriestype=:shape, color=:bisque3, linewidth=0.5, kwargs...)

Plot a coastline with longitude and latitude axes (i.e. without a map projection).

The aspect_ratio of the plot is set as the reciprocal of the mean of the ylims values, to preserve shapes near that spot.

Arguments

coastline the coastline, as constructed using coastline or (less

commonly) Coastline.

xlims and ylims control the ranges of the longitude and latitude axes,

respectively.

seriestype, color and linewidth control the rendering of land regions.

These values are passed to the base-level plot function; for details, see the documentation provided by the Plots package.

OceanAnalysis.plot_coastline! — Function

plot_coastline!(coastline::Coastline;
    seriestype=:shape, color=:bisque3, linewidth::Real=0.5)

Add a coastline to an existing plot.

This shares several arguments with plot_coastline, but not those that could alter the geometry. Note that the plot limits are inherited from the existing plot, so xlim and ylim should not be supplied in the kwargs... grouping.

Arguments

coastline the coastline, as constructed using coastline or, by

more advanced users, using Coastline.

seriestype, color and linewidth control the rendering of land regions.

These values are passed to the base-level plot function; for details, see the documentation provided by the Plots package.

OceanAnalysis.plot_profile — Function

plot_profile(ctd::Ctd; which::String="CT", vertical::String="pressure",
    abbreviate::Bool=false, legend::Bool=false, tickfontsize=8, tickdirection=:out,
    labelfontsize=8, debug::Int64=0, kwargs...)

Plot an oceanographic profile for data contained in ctd, showing how the variable named by which depends on pressure. The variable is drawn on the x axis and pressure on the y axis. Following oceanographic convention, pressure increases downwards on the page and the "x" axis is drawn at the top. The permitted values of which are "CT" for the Gibbs Seawater formulation of Conservative Temperature, "N2" for N², the square of the buoyancy frequency, "SA" for the Gibbs Seawater formulation of Absolute Salinity, "salinity" for Practical Salinity, "sigma0" for the Gibbs Seawater formulation of density anomaly referenced to the surface, "spiciness0" for the Gibbs Seawater seawater spiciness referenced to the surface, and "temperature" for in-situ temperature.

The default Julia font sizes on axes are overridden in this function, with 8-point being used for both the numbers on axes (tickfontize) and the names of axes (labelfontsize). (The tickfontsize matches the Julia default, but the labelfontsize is smaller than the Julia default. The idea is to not waste space by using fonts that are larger than what journals require.)

The kwargs... argument is used for arguments to be sent to plot(). For example, the default way to display the profile diagram is constructed with a blue line connecting points, but using e.g.

plot_profile(ctd, which="SA", seriestype=:scatter, seriescolor=:red)

yields red-filled circles, instead; see https://docs.juliaplots.org/stable/ for more on the many plotting controls available in Julia.

See also the plot_TS function.

Examples

using OceanAnalysis, Plots
# Read an Argo file
pkgdir = dirname(dirname(pathof(OceanAnalysis)))
f = joinpath(pkgdir, "data", "D4902911_095.nc")
d = read_argo(f, 1);
# Plot profiles of Conservative Temperature, Absolute Salinity, and potential
# density anomaly with respect to surface pressure.
p1 = plot_profile(d, which="CT")
p2 = plot_profile(d, which="SA")
p3 = plot_profile(d, which="sigma0")
plot(p1, p2, p3, layout=(1, 3), size=(800, 400))

OceanAnalysis.plot_TS — Function

plot_TS(ctd::Ctd; sigma0_levels=[], spiciness0_levels=0,
    draw_freezing=true, abbreviate=false,
    framestyle=:box, color=:black, seriestype=:scatter, markersize=2,
    legend=false, gridstyle=:dash, tickfontsize=8, tickdirection=:out,
    labelfontsize=8, debug::Int64=0, kwargs...)

Plot an oceanographic TS diagram, with the Gibbs Seawater equation of state.

Whether contours of density and spiciness are drawn depends on values of the sigma0_levels and spiciness0_level arguments. There are 4 categories. (1) Setting either to 0 prevents contouring. (2) Setting either to [] enables contours with automatic selection of levels. (3) Setting either to a positive integer provides a suggestion for the number of levels, with the actual number being set by pretty), which is provided with the integer. (4) Setting either to a vector of numbers specifies those numbers as the levels.

By default, a freezing-point line is drawn (if it is within the range of the data). This behaviour is skipped if draw_freezing is false.

By default, axis names are written in long form; set abbreviate=true for shorter versions.

Information about the analysis is printed if debug is set to true.

Apart from that, the other parameters have the usual meanings for Julia plots. For example, color is set to black, to override the Julia default, etc. In addition to those parameters, the kwargs... argument represents any other argument that is accepted by plot. This is illustrated in the example, which a title is added to the plot for a built-in CNV-formatted CTD file.

using OceanAnalysis, Plots, Dates
pkgdir = dirname(dirname(pathof(OceanAnalysis)))
f = joinpath(pkgdir, "data", "ctd.cnv")
ctd = read_ctd_cnv(f);
plot_TS(ctd, title="Built-in CTD file", titlefontsize=9)

See also plot_profile.

OceanAnalysis.pressure_from_depth — Function

Compute sea pressure (dbar) from depth (m) and latitude (deg).

Examples

julia> using OceanAnalysis

julia> pressure_from_depth(10.0)
10.082069761243858

OceanAnalysis.pressure_from_z — Function

Compute sea pressure (dbar) from vertical coordinate (m) and latitude (deg).

Examples

julia> using OceanAnalysis

julia> pressure_from_z(-10.0)
10.082069761243858

OceanAnalysis.pretty — Function

pretty(x, n::Int64=5; debug::Int64=0)

Calculate sub-intervals with 125 scaling

This function finds intervals that are multiples of 1, 2 or 5. The results are useful for contour intervals, because the built-in contour() function (Reference 1) simply divides the range into equal intervals, so that labelled contours can be quite ugly.

Examples

julia> using OceanAnalysis

julia> pretty([22.299, 25.091])
8-element Vector{Float64}:
 22.0
 22.5
 23.0
 23.5
 24.0
 24.5
 25.0
 25.5

References

https://github.com/JuliaGeometry/Contour.jl/blob/daad6eb0b1464dbc7e824bf8384cad54a3b76445/src/Contour.jl#L100)

OceanAnalysis.read_amsr — Function

read_amsr(filename::String, field::String="SST", debug=0)

Reads a NetCDF file containing AMSR data.

This returns a value of the Amsr type, with metadata containing the filename along with vectors holding the longitude and latitude of the grid. The data field holds a matrix of the data element with the indicated name (e.g. name="SST" for sea-surface temperature).

Arguments

filename a string indicating the location of the local file.

field a string used to identify the data field to be extracted. If

field="?" then read_amsr returns a vector of strings containing extractable data. Otherwise, if field names one of those items, then read_amsr returns that dataset.

Examples

using OceanAnalysis, Plots
f = "~/data/amsr/RSS_AMSR2_ocean_L3_3day_2025-09-07_v08.2.nc";
d = read_amsr(f, "SST");
longitude = d.metadata["longitude"];
latitude = d.metadata["latitude"];
SST = d.data;
heatmap(longitude, latitude, SST, framestyle=:box, aspect_ratio=:equal,
    xlims=(0, 360), ylims=(-90, 90), dpi=300, size=(800, 400),
    title=f * ": SST", titlefontsize=9)

OceanAnalysis.read_argo — Function

read_argo(filename::String; column::Int64=1, add_teos::Bool=true,
    require_valid::Bool=true, debug::Int64=0)

Read an Argo file and return a Ctd object that holds salinity, temperature, pressure (and derived columns) but no other columns from the file. This function is in an early stage of development; please report problems as issues on <www.github.com/dankelley/OceanAnalysis.jl/issues>.

The value of add_teos is passed to as_ctd, where it indicates whether to add TEOS-10 variables such as SA, CT, sigma0 and spiciness0 to the data portion of the return value.

If require_valid is true (the default) then an error is reported if the file lacks one of three required data columns, or either longitude or latitude. An error is also reported if any of these items consists entirely of missing values. This is because such files are unlikely to be of much use. In some cases, setting require_valid to false may permit the file to be read, but this has not been tested, since the results in such cases are not likely to be of practical use.

Set debug to a positive integer to cause read_argo() to print messages during processing. This can be handy if problems arise.

Return value

The read_argo() function returns a Ctd object that has two components, a Dict named .metadata and DataFrame named .data. The .metadata entries are named "cycle", "data_mode", "date_creation", "filename", "latitude", "longitude", "platform", and "time". The .data columns are named "pressure", "salinity" and "temperature", as copied from fields in the NetCDF file named "PRES", "PSAL" and "TEMP"; no other NetCDF fields are copied in this version of read_argo().

Examples

julia> using OceanAnalysis, Plots

julia> pkgdir = dirname(dirname(pathof(OceanAnalysis)));

julia> f = joinpath(pkgdir, "data", "D4902911_095.nc");

julia> d = read_argo(f);

julia> d.metadata["time"]
2019-10-14T23:43:44.003

julia> d.metadata["latitude"]
40.45216

julia> d.metadata["longitude"]
-66.38298

julia> first(d.data,3)
3×7 DataFrame
 Row │ salinity  temperature  pressure  SA       CT       sigma0   spiciness0
     │ Float64   Float64      Float64   Float64  Float64  Float64  Float64
─────┼────────────────────────────────────────────────────────────────────────
   1 │   34.913       19.513      0.48  35.0786  19.5079  24.8272     3.31464
   2 │   34.91        19.527      1.0   35.0756  19.5219  24.8213     3.31603
   3 │   34.912       19.524      2.0   35.0776  19.5187  24.8237     3.31669

OceanAnalysis.read_argo_index — Function

read_argo_index(file::String; trim::Bool=true, header::Int64=9, debug::Int64=0)

Read a file downloaded by get_argo_index.

This relies on there being exactly header lines of header, the last of which names the columns. The default value of 9 works with index files downloaded from the ifremer.fr server, as of 2025-09-08.

First, the date column is converted to a DateTime column named time. If trim is true, then the original date column is removed, along with the the columns named institution, date_update, ocean, and profiler_type.

Return

read_argo_index returns a DataFrame with column names "file", "latitude", "longitude", and "time". Note that the "file" column holds information on the location on remote servers, as is required for use as the file argument of get_argo_file.

OceanAnalysis.read_ctd_cnv — Function

read_ctd_cnv(filename; add_teos=false, debug=0)

Read a Seabird CTD file in cnv format, optionally adding TEOS-10 variables.

Returns a Ctd object that holds metadata (a Dict) and data (a DataFrame). metadata item holds header (a vector of strings, one per line from the start down to a line containing #END), plus some particular items scanned from that header. data holds the columnar data read from the file, along with renamed values in standard nomenclature. At present, the only renamed items are salinity, temperature, and pressure. Note that if the data file indicates temperature is on the T68 scale, then this is converted to the standard modern scale, T90, before saving as temperature.

A message is printed if no data in the file are labelled with names that are recognized as salinity, temperature, or pressure, because these quantities are required for any meaningful CTD dataset.

The value of add_teos is passed to as_ctd, where it indicates whether to add TEOS-10 variables such as SA, CT, sigma0 and spiciness0 to the data portion of the return value.

Examples

julia> using OceanAnalysis

julia> f = joinpath(dirname(dirname(pathof(OceanAnalysis))), "data", "ctd.cnv");

julia> d = read_ctd_cnv(f, add_teos=false);

julia> d.metadata["time"]
2003-10-15T11:38:38

julia> d.metadata["latitude"]
44.684266666666666

julia> d.metadata["longitude"]
-63.643883333333335

julia> names(d.data)
10-element Vector{String}:
 "salinity"
 "temperature"
 "pressure"
 "scan"
 "timeS"
 "pr"
 "depS"
 "t090"
 "sal00"
 "flag"

OceanAnalysis.SA — Function

SA(salinity, pressure, longitude, latitude)
SA(ctd)

Compute Absolute Salinity (SA), using gsw_sa_from_sp() in the GibbsSeaWater package.

The first form takes single values and returns a single value. The second form extracts values from a Ctd object and then calls the first form as SA.() so that it returns a vector of SA values.

Examples

julia> using OceanAnalysis

julia> SA(35.0, 100.0, -30.0, 30.0)
35.165308620244

OceanAnalysis.T90_from_T48 — Function

T90 = T90_from_T48(T48::Float64)

Convert a temperature from the T48 scale to the T90 scale.

See also T90_from_T68.

Examples

julia> using OceanAnalysis

julia> T90_from_T48(10.0)
9.993641526033752

OceanAnalysis.T90_from_T68 — Function

T90 = T90_from_T68(T68::Float64)

Convert a temperature from the T68 scale to the T90 scale.

See also T90_from_T48.

Examples

julia> using OceanAnalysis

julia> T90_from_T68(10.0)
9.997600575861792

OceanAnalysis.z_from_pressure — Function

Compute vertical coordinate (m) from sea pressure (dbar)

See also depth_from_pressure.

Examples

julia> using OceanAnalysis

julia> z_from_pressure(100.0)
-99.16434938694897