Constants and Data types

Constants:

OceanAnalysis.GSW_INVALID_THRESHOLD — Constant

GSW_INVALID_THRESHOLD

This value, 1.0e15, is returned by several functions in the GibbsSeaWater package to indicate an unphysical result of a computation of e.g. Absolute Salinity, etc. The fix_gsw_bad_code and fix_gsw_bad_code! functions provide a convenient way to replace any such values with NaN.

OceanAnalysis.DEFAULT_LONGITUDE — Constant

DEFAULT_LONGITUDE

This value, -30.0, is used as a default by SA, if the call omits the longitude argument.

OceanAnalysis.DEFAULT_LATITUDE — Constant

DEFAULT_LATITUDE

This value, 45.0, is used as a default by SA, if the call omits the longitude argument.

Data types:

OceanAnalysis.OA — Type

Base type in the OceanAnalysis package.

This is an abstract type, from which other types in the library are derived.

OceanAnalysis.Adp — Type

A type to hold acoustic-Doppler profiler data

Adp is a type used to store data from an ADP (acoustic-Doppler profiler). It holds two items: data (a Dict containing a set of vectors and arrays that hold information retrieved from the file, e.g. "velocity" is a 3D array with inferred velocity) and metadata (a Dict with information about the data).

Adp objects may be created with read_adp_rdi, which only handles RDI data. (Use the oce R package if you need other instrument types.)

See ?read_adp_rdi for information on reading such files, and see the online documentation for an illustration.

OceanAnalysis.Amsr — Type

A type to hold AMSR satellite data

This holds AMSR satellite data as read by read_amsr.

The metadata element is a Dict that holds the source filename, the sensor, the name of the stored variable, the observation interval ("time_coverage_start" and time_coverage_end) and the vectors (longitude and latitude) that define the grid. The data element is a matrix holding the gridded data.

See the online documentation for an illustration.

OceanAnalysis.Argo — Type

A type to hold Argo data

Argo is a type used to store data from Argo floats.

Argo objects hold two components: a DataFrame named data that holds the actual data, and a Dict named metadata that stores information about the data, e.g. the sampling time, the location, etc. Objects of type Argo are returned by read_argo, which reads the NetCDF files in which Argo data are distributed.

For many purposes, it may be useful to convert Argo objects into Ctd objects, using as_ctd.

See the online documentation for an illustration.

OceanAnalysis.Ctd — Type

A type to hold CTD data

Ctd is a type used to store data from CTD instruments.

Ctd objects hold two items: (1) a DataFrame named data that holds the actual data, including pressure, salinity, and temperature, perhaps along with other data and (2) a Dict named metadata that stores information about the data, such as the time of observation and the latitude and longitude at which the observation was made.

Objects of type Ctd are returned by as_ctd, read_ctd_cnv, read_ctd_rsk and read_argo. Such objects can be passed to plotting functions plot_profile and plot_TS, and to some functions relating to seawater properties, such as SA and other TEOS-10 related functions, as well as functions relating to the distributions of such properties, such as N2.

See the online documentation for an illustration.

OceanAnalysis.Coastline — Type

A type to hold coastline data

Coastline is a type to hold coastline data. Its metadata element is a Dict that may hold the source filename or other information (or may be empty). Its data element is a DataFrame holding columns named longitude and latitude, with NaN values separating continents and/or elements within them such as countries or subregions of countries.

See the online documentation for an illustration.

OceanAnalysis.Dem — Type

A type to hold DEM (digital-elevation model) data

Dem is a type used to store DEM (digital-elevation model) data. It holds two items: data (a Dict that holds vectors x, y, longitude, latitude and corresponding matrix z) and metadata (a Dict that holds information about the dataset).

Dem objects may be created with read_dem. Subsets may be taken with subset_dem](@ref). Plots may be made with plot_dem@ref).

OceanAnalysis.Echosounder — Type

A type to hold echosounder data

Echosounder is a type to hold data from a Biosonics scientific echosounder, as read with read_echosounder. Its metadata and data are both Dict objects.

See the online documentation for an illustration.

OceanAnalysis.Glider — Type

A type to hold glider data

Glider is a type used to store data from ocean gliders. It holds two items: data (a DataFrame that holds data in array form) and metadata (a Dict with information about the data).

Glider objects may be created with read_glider.

OceanAnalysis.Nonna — Type

A type to hold NONNA bathymetry data

This holds bathymetric data as read by read_nonna.

The metadata element is a Dict that holds the source filename along with vectors named longitude and latitude that define the grid. The data element stores a matrix of bathymetry data in terms of height above mean sea level, in metres.

See the online documentation for an illustration.

OceanAnalysis.Section — Type

Section is a type that holds section data

The data portion of a Section object is a Vector of Ctd objects, while the metadata portion may contain information about the collection.

Section objects may be assembled from CTD objects using as_section or read from CTD files with read_section. They may be gridded with grid_section and plotted with plot_section.

See the online documentation for an illustration.

OceanAnalysis.Topography — Type

A type to hold topography data

This holds topography data as read by read_topography.

The metadata element is a Dict that holds the source filename along with vectors holding the longitude and latitude values that define the grid.

The data element stores a matrix of topography data in terms of height above mean sea level, in metres.

See the online documentation for an illustration.